Code cleanup in Reqnroll.Formatters.PayloadProcessing
