+++ Reqnroll/Formatters/ExecutionTracking/HookStepDefinition.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using Reqnroll.Formatters.PayloadProcessing.Cucumber;
 
 namespace Reqnroll.Formatters.ExecutionTracking
 {
