+++ Reqnroll/Formatters/PayloadProcessing/Cucumber/ICucumberMessageVisitor.cs
@@ -2,9 +2,10 @@
 
 namespace Reqnroll.Formatters.PayloadProcessing.Cucumber;
 
-// This interface is used to support the implementation of an External Vistor pattern against the Cucumber Messages.
-// Visitors impmlement this interface and then invoke it using the helper class below.
-
+/// <summary>
+/// This interface is used to support the implementation of an external visitor pattern against the Cucumber Messages.
+/// Visitors implement this interface and then invoke it using the helper class below.
+/// </summary>
 public interface ICucumberMessageVisitor
 {
     // Existing methods
